requirements.txt
numpy
pandas
segmentation-models-pytorch==0.2.0
albumentations
#GDAL==2.3.3
seaborn
opencv-python
Pillow
torchvision>=0.5.0
pretrainedmodels==0.7.4
efficientnet-pytorch==0.6.3
timm==0.4.12
tqdm